Split index.js code into multiple files
