+++ javascript/inventory/inventory.js
@@ -1,0 +1,13 @@
+class Inventory {
+  constructor(items) {
+    this.items = items;
+  }
+
+  updatePrice() {
+    this.items.forEach((item) => {
+      item.updatePriceAndSellBy();
+    });
+  }
+}
+
+module.exports = Inventory;
